PlantStatusDetection/requirements.txt
Flask
gunicorn
Pillow
google-generativeai
Werkzeug